requirement.txt
matplotlib.pyplot
fuzzywuzzy
sqlalchemy
